reorganized menu logic
